Refactor Search/ListablePodcast
Move html and logic for subscriptions to new component (SubscribeButton)
Move html for rating and mor options to separet components
Move styling for corresponding components to separet files
