+++ src/Assets/Styles/Components/Common/more-options-button.scss
@@ -1,0 +1,7 @@
+.more-options-button {
+  background-image: url(../Icons/more.svg);
+  background-repeat: no-repeat;
+  background-size: contain;
+  width: 70px;
+  height: 70px;
+}
