+++ src/Assets/Styles/Components/Common/rating.scss
@@ -1,0 +1,12 @@
+.rating {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 0;
+
+  img {
+    width: 40px;
+    height: 40px;
+  }
+}
