+++ src/Assets/Styles/Components/Common/subscribe-button.scss
@@ -1,0 +1,24 @@
+.subscribe-button {
+  background-color: #78d69f;
+  color: #2e2929;
+  font-size: 1.5em;
+  border-radius: 4px;
+  padding: 0.25em 0.5em;
+  height: 35px;
+  line-height: 0.9em;
+  box-shadow: 4px 4px 0 rgba(0, 0, 0, 0.25);
+  max-width: 128px;
+
+  &.unsubscribe {
+    background-color: #f6c821;
+    font-size: 1.25em;
+  }
+
+  &.loading {
+    background-image: url(../Icons/loader.svg);
+    background-size: 20px;
+    background-repeat: no-repeat;
+    background-color: #56ccf2;
+    background-position: 50%;
+  }
+}
