+++ src/Assets/Styles/Components/Search/listable-podcast-searchresult.scss
@@ -53,50 +53,4 @@
     align-items: flex-start;
     flex-direction: column;
   }
-
-  .rating {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin: 0;
-
-    img {
-      width: 40px;
-      height: 40px;
-    }
-  }
-
-  .more-options-button {
-    background-image: url(../Icons/more.svg);
-    background-repeat: no-repeat;
-    background-size: contain;
-    width: 70px;
-    height: 70px;
-  }
-
-  .subscribe-button {
-    background-color: #78d69f;
-    color: #2e2929;
-    font-size: 1.5em;
-    border-radius: 4px;
-    padding: 0.25em 0.5em;
-    height: 35px;
-    line-height: 0.9em;
-    box-shadow: 4px 4px 0 rgba(0, 0, 0, 0.25);
-    max-width: 128px;
-
-    &.unsubscribe {
-      background-color: #f6c821;
-      font-size: 1.25em;
-    }
-
-    &.loading {
-      background-image: url(../Icons/loader.svg);
-      background-size: 20px;
-      background-repeat: no-repeat;
-      background-color: #56ccf2;
-      background-position: 50%;
-    }
-  }
 }
