+++ src/Assets/Styles/index.scss
@@ -11,7 +11,10 @@
 @import './Components/Playback/playback-interface.scss';
 @import './Components/Playback/progress-bar.scss';
 @import './Components/Auth/register-form.scss';
-@import './Components/Layout/modal.scss';
+@import './Components/Common/subscribe-button.scss';
+@import './Components/Common/modal.scss';
+@import './Components/Common/rating.scss';
+@import './Components/Common/more-options-button.scss';
 @import './Components/Auth/login-form.scss';
 @import './Components/Layout/menu.scss';
 @import './Components/Search/search-controlles.scss';
