+++ src/Components/Common/MoreOptionsButton.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function MoreOptionsButton(): JSX.Element {
+  return (
+    <button type="button" aria-label="more-options-button" className="more-options-button" />
+  );
+}
+
+export default MoreOptionsButton;
