+++ src/Components/Search/SearchResults.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, ComponentType } from 'react';
 import List from '../Common/List';
-import ListablePodcast from '../../Containers/Search/ListablePodcast';
+import ListablePodcast from './ListablePodcast';
 import ListableEpisode from '../../Containers/Search/ListableEpisode';
 import ListableUser from './ListableUser';
 import Loader from '../Layout/Loader';
