+++ src/Components/Subscription/Category.tsx
@@ -2,7 +2,7 @@ import React, {
   useState, useEffect, useMemo,
 } from 'react';
 import List from '../Common/List';
-import ListablePodcast from '../../Containers/Search/ListablePodcast';
+import ListablePodcast from '../Search/ListablePodcast';
 import { Podcast } from '../../Models/Podcast';
 import { Category } from '../../Models/Category';
